+++ src/main/java/com/twilio/rest/video/v1/CompositionCreator.java
@@ -52,7 +52,7 @@ public class CompositionCreator extends Creator<Composition> {
     /**
      * An object that describes the video layout of the composition in terms of
      * regions. See [Specifying Video
-     * Layouts](https://www.twilio.com/docs/video/api/compositions-resource#specifying-video-layouts) for more info..
+     * Layouts](https://www.twilio.com/docs/video/api/compositions-resource#specifying-video-layouts) for more info. Please, be aware that either video_layout or audio_sources have to be provided to get a valid creation request.
      *
      * @param videoLayout An object that describes the video layout of the
      *                    composition
@@ -70,7 +70,8 @@ public class CompositionCreator extends Creator<Composition> {
      * specified in `audio_sources_excluded`. The track names in this parameter can
      * include an asterisk as a wild card character, which will match zero or more
      * characters in a track name. For example, `student*` includes `student` as
-     * well as `studentTeam`..
+     * well as `studentTeam`. Please, be aware that either video_layout or
+     * audio_sources have to be provided to get a valid creation request.
      *
      * @param audioSources An array of track names from the same group room to merge
      * @return this
@@ -87,7 +88,8 @@ public class CompositionCreator extends Creator<Composition> {
      * specified in `audio_sources_excluded`. The track names in this parameter can
      * include an asterisk as a wild card character, which will match zero or more
      * characters in a track name. For example, `student*` includes `student` as
-     * well as `studentTeam`..
+     * well as `studentTeam`. Please, be aware that either video_layout or
+     * audio_sources have to be provided to get a valid creation request.
      *
      * @param audioSources An array of track names from the same group room to merge
      * @return this